complete implementation of sorting items with configuration
